chore(tauri/ui): move walletconnect store fns into their own file
