+++ tauri-app/src/lib/components/ModalAddOrder.svelte
@@ -3,7 +3,8 @@
   import ButtonLoading from '$lib/components/ButtonLoading.svelte';
   import { walletDerivationIndex, walletAddress } from '$lib/stores/wallets';
   import InputLedgerWallet from './InputLedgerWallet.svelte';
-  import { walletconnectModal, walletconnectAccount, orderbookAddress } from '$lib/stores/settings';
+  import { orderbookAddress } from '$lib/stores/settings';
+  import { walletconnectModal, walletconnectAccount } from '$lib/stores/walletconnect';
   import type { Deployment } from '$lib/typeshare/config';
   import { orderAdd, orderAddCalldata } from '$lib/services/order';
   import { ethersExecute } from '$lib/services/ethersTx';
